requirements.txt
Pillow
graphene-django>=2.0
django-filter>=2
Faker
factory_boy
coverage
